Improve a bit the back-end for users
Add the ability to filter the users by distribution point.

Also, allow the ordering of their name. We have to remove the `full_name`
column and rely on both `first_name` and `last_name` as Active Admin
will only allow ordering through SQL.

Finally, upgrade the Ruby version to 2.4.1 as the block syntax is
sometimes buggy on 2.3.